Move code to BenchmarkToolsExt
