+++ benchmarks/benchmarks.jl
@@ -2,23 +2,42 @@ using Pkg
 # To ensure we benchmark the local version of DynamicPPL, dev the folder above.
 Pkg.develop(; path=joinpath(@__DIR__, ".."))
 
-using DynamicPPLBenchmarks: Models, make_suite, model_dimension
+using DynamicPPL: DynamicPPL, make_benchmark_suite, VarInfo
 using BenchmarkTools: @benchmark, median, run
 using PrettyTables: PrettyTables, ft_printf
+using ForwardDiff: ForwardDiff
+using Mooncake: Mooncake
+using ReverseDiff: ReverseDiff
 using StableRNGs: StableRNG
 
-rng = StableRNG(23)
+include("Models.jl")
+
+"""
+    model_dimension(model, islinked)
+
+Return the dimension of `model`, accounting for linking, if any.
+"""
+function model_dimension(model, islinked)
+    vi = VarInfo()
+    model(vi)
+    if islinked
+        vi = DynamicPPL.link(vi, model)
+    end
+    return length(vi[:])
+end
 
 # Create DynamicPPL.Model instances to run benchmarks on.
-smorgasbord_instance = Models.smorgasbord(randn(rng, 100), randn(rng, 100))
+smorgasbord_instance = Models.smorgasbord(
+    randn(StableRNG(23), 100), randn(StableRNG(23), 100)
+)
 loop_univariate1k, multivariate1k = begin
-    data_1k = randn(rng, 1_000)
+    data_1k = randn(StableRNG(23), 1_000)
     loop = Models.loop_univariate(length(data_1k)) | (; o=data_1k)
     multi = Models.multivariate(length(data_1k)) | (; o=data_1k)
     loop, multi
 end
 loop_univariate10k, multivariate10k = begin
-    data_10k = randn(rng, 10_000)
+    data_10k = randn(StableRNG(23), 10_000)
     loop = Models.loop_univariate(length(data_10k)) | (; o=data_10k)
     multi = Models.multivariate(length(data_10k)) | (; o=data_10k)
     loop, multi
@@ -34,7 +53,7 @@ end
 chosen_combinations = [
     (
         "Simple assume observe",
-        Models.simple_assume_observe(randn(rng)),
+        Models.simple_assume_observe(randn(StableRNG(23))),
         :typed,
         :forwarddiff,
         false,
@@ -50,14 +69,14 @@ chosen_combinations = [
     ("Loop univariate 10k", loop_univariate10k, :typed, :mooncake, true),
     ("Multivariate 10k", multivariate10k, :typed, :mooncake, true),
     ("Dynamic", Models.dynamic(), :typed, :mooncake, true),
-    ("Submodel", Models.parent(randn(rng)), :typed, :mooncake, true),
+    ("Submodel", Models.parent(randn(StableRNG(23))), :typed, :mooncake, true),
     ("LDA", lda_instance, :typed, :reversediff, true),
 ]
 
 # Time running a model-like function that does not use DynamicPPL, as a reference point.
 # Eval timings will be relative to this.
 reference_time = begin
-    obs = randn(rng)
+    obs = randn(StableRNG(23))
     median(@benchmark Models.simple_assume_observe_non_model(obs)).time
 end
 
@@ -65,7 +84,7 @@ results_table = Tuple{String,Int,String,String,Bool,Float64,Float64}[]
 
 for (model_name, model, varinfo_choice, adbackend, islinked) in chosen_combinations
     @info "Running benchmark for $model_name"
-    suite = make_suite(model, varinfo_choice, adbackend, islinked)
+    suite = make_benchmark_suite(StableRNG(23), model, varinfo_choice, adbackend, islinked)
     results = run(suite)
     eval_time = median(results["evaluation"]).time
     relative_eval_time = eval_time / reference_time
